chore: use long instead of int for topic sequence no. and page

+++ momento-sdk/src/main/java/momento/sdk/internal/SubscriptionState.java
@@ -4,8 +4,8 @@ package momento.sdk.internal;
 public class SubscriptionState {
 
   private Runnable unsubscribeFn;
-  private Integer lastTopicSequenceNumber;
-  private Integer lastTopicSequencePage;
+  private Long lastTopicSequenceNumber;
+  private Long lastTopicSequencePage;
   private boolean isSubscribed;
 
   /** Constructs a new SubscriptionState instance with default values. */
@@ -19,12 +19,12 @@ public class SubscriptionState {
    *
    * @return The topic sequence number to resume from.
    */
-  public int getResumeAtTopicSequenceNumber() {
+  public long getResumeAtTopicSequenceNumber() {
     return (lastTopicSequenceNumber != null ? lastTopicSequenceNumber : -1) + 1;
   }
 
   /** Sets the topic sequence number to resume the subscription from. */
-  public void setResumeAtTopicSequenceNumber(int lastTopicSequenceNumber) {
+  public void setResumeAtTopicSequenceNumber(long lastTopicSequenceNumber) {
     this.lastTopicSequenceNumber = lastTopicSequenceNumber;
   }
 
@@ -33,12 +33,12 @@ public class SubscriptionState {
    *
    * @return The topic sequence page to resume from.
    */
-  public int getResumeAtTopicSequencePage() {
+  public long getResumeAtTopicSequencePage() {
     return lastTopicSequencePage != null ? lastTopicSequencePage : 0;
   }
 
   /** Sets the topic sequence page to resume the subscription from. */
-  public void setResumeAtTopicSequencePage(int lastTopicSequencePage) {
+  public void setResumeAtTopicSequencePage(long lastTopicSequencePage) {
     this.lastTopicSequencePage = lastTopicSequencePage;
   }
